"refactor : user 리팩토링"
