+++ src/main/java/com/example/hipreader/domain/user/dto/request/UpdateUserRequestDto.java
@@ -1,6 +1,6 @@
 package com.example.hipreader.domain.user.dto.request;
 
-import com.example.hipreader.domain.user.gender.Gender;
+import com.example.hipreader.domain.user.vo.Gender;
 
 import lombok.Getter;
