+++ src/main/java/com/example/hipreader/domain/user/vo/UserRole.java
@@ -1,5 +1,4 @@
-package com.example.hipreader.domain.user.role;
-
+package com.example.hipreader.domain.user.vo;
 
 import java.util.Arrays;
