+++ src/main/java/com/example/hipreader/domain/userbook/repository/UserBookRepositoryCustom.java
@@ -7,7 +7,7 @@ import org.springframework.data.domain.Pageable;
 
 import com.example.hipreader.domain.book.entity.Book;
 import com.example.hipreader.domain.book.genre.Genre;
-import com.example.hipreader.domain.user.gender.Gender;
+import com.example.hipreader.domain.user.vo.Gender;
 
 public interface UserBookRepositoryCustom {
 	Page<Book> findRecommendedBooks(Integer age, Gender gender, String categoryName, Pageable pageable);
