+++ src/main/java/com/example/hipreader/domain/userbook/repository/UserBookRepositoryImpl.java
@@ -13,7 +13,7 @@ import org.springframework.stereotype.Repository;
 
 import com.example.hipreader.domain.book.entity.Book;
 import com.example.hipreader.domain.book.genre.Genre;
-import com.example.hipreader.domain.user.gender.Gender;
+import com.example.hipreader.domain.user.vo.Gender;
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.jpa.impl.JPAQueryFactory;
